+++ android/app/src/main/java/io/coursepick/coursepick/presentation/routetfinder/RouteFinderChoiceDialogFragment.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.DialogFragment
 import io.coursepick.coursepick.databinding.DialogRouteFinderChoiceBinding
 import io.coursepick.coursepick.domain.Coordinate
-import io.coursepick.coursepick.presentation.getSerializableCompat
+import io.coursepick.coursepick.presentation.compat.getSerializableCompat
 import io.coursepick.coursepick.presentation.preference.CoursePickPreferences
 
 class RouteFinderChoiceDialogFragment :
